requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pytorch-lightning>=2.1.0
efficientnet-pytorch>=0.7.1
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pandas>=2.2.0
scikit-learn>=1.4.0
albumentations>=1.3.1
pyyaml>=6.0.1
sentry-sdk>=1.40.0
opencv-python-headless>=4.9.0
matplotlib>=3.8.0
tensorboard>=2.15.0
